Start with CNN; pickling some models to disk; bug fixes
Added word2vec support and embedding matrix.  Embedding matrix is
pickled to disk and reloaded if present; the same is done to the IMDB
movie review dataset (which is very slow to load).

CNN model is built, but not yet trained.

execute.py came with some bugs and warts that were fixed; some string
checks were incorrect between the documentation and the code, and I
added an error check so that mismatched strings at least trigger an
error rather than failing cryptically.